reiterate instructions for exiting repl
When trying things on my own, I often gloss over code blocks esp. in a
situation like this where it's just a confirmation. Getting stuck in a
repl is a terrible experience, so I think it will be nice to have a
little reminder on how to exit in case people aren't reading every word.