Simplify import section in Getting Started
Feedback from using this in a course: The old import example of Dict is confusing because all of a sudden you have two type varaibels k and v in your type signatures etc. If all you want to show are imports, something simple like Bitwise is maybe an easier example.